core/src/main/java/ru/mobnius/core/utils/HardwareUtil.java
package ru.mobnius.core.utils;

import android.Manifest;
import android.annotation.SuppressLint;
import android.content.Context;
import android.content.Intent;
import android.content.IntentFilter;
import android.content.pm.PackageManager;
import android.os.BatteryManager;
import android.os.Build;
import android.provider.Settings;

import androidx.core.app.ActivityCompat;

public class HardwareUtil {
    /**
     * Возвращается уникальный номер, либо модель если нет доступа
     * @param context activity
     * @return Возвращается уникальный номер
     */
    @SuppressLint("HardwareIds")
    public static String getNumber(Context context) {
        if (ActivityCompat.checkSelfPermission(context, Manifest.permission.READ_PHONE_STATE) != PackageManager.PERMISSION_GRANTED) {
            return Settings.Secure.getString(context.getContentResolver(),
                    Settings.Secure.ANDROID_ID);
        }
        return Build.MODEL;
    }

    /**
     * Заряд батареи
     * @param context контекст
     * @return возвращается заряд батареи в процентах
     */
    public static Integer getBatteryPercentage(Context context) {

        IntentFilter iFilter = new IntentFilter(Intent.ACTION_BATTERY_CHANGED);
        Intent batteryStatus = context.registerReceiver(null, iFilter);

        int level = batteryStatus != null ? batteryStatus.getIntExtra(BatteryManager.EXTRA_LEVEL, -1) : -1;
        int scale = batteryStatus != null ? batteryStatus.getIntExtra(BatteryManager.EXTRA_SCALE, -1) : -1;

        float batteryPct = level / (float) scale;

        return (int) (batteryPct * 100);
    }
}
